refactor: 단위 테스트 작성에 따른 BasicAcceptanceTest 이동
