+++ src/test/java/roomescape/acceptance/ReservationTimeAcceptanceTest.java
@@ -15,6 +15,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.DynamicTest;
 import org.junit.jupiter.api.TestFactory;
 import org.springframework.test.context.jdbc.Sql;
+import roomescape.BasicAcceptanceTest;
 import roomescape.controller.response.AvailableReservationTimeResponse;
 
 class ReservationTimeAcceptanceTest extends BasicAcceptanceTest {
