+++ src/test/java/roomescape/acceptance/ThemeAcceptanceTest.java
@@ -16,6 +16,7 @@ import org.junit.jupiter.api.TestFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.test.context.jdbc.Sql;
+import roomescape.BasicAcceptanceTest;
 import roomescape.TestFixtures;
 import roomescape.controller.response.ThemeResponse;
